Move today summary calculation into domain use case
